requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
numpy>=1.26.0
sentence-transformers==2.2.2
qdrant-client==1.7.0 